UM-ZAD1-KNN/requirements.txt
numpy==2.3.4
pandas==2.3.3
tmdbsimple==2.9.1
python-dotenv==1.1.1
scikit-learn==1.7.2